requirements-lock.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anaconda-anon-usage @ file:///C:/b/abs_1az9wuczyu/croot/anaconda-anon-usage_1749054795139/work
annotated-types==0.7.0
archspec @ file:///croot/archspec_1709217642129/work
arrow==1.3.0
attrs==25.3.0
Authlib==1.6.5
bandit==1.7.7
bar-directory-recon==0.1.1
beautifulsoup4==4.14.2
black==24.3.0
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
boolean.py==5.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
build==1.3.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
conda @ file:///C:/b/abs_78fae1xfc9/croot/conda_1749171262913/work/conda-src
conda-anaconda-telemetry @ file:///C:/b/abs_d4qxn0oeix/croot/conda-anaconda-telemetry_1751062015410/work
conda-anaconda-tos @ file:///C:/b/abs_4csynbis2x/croot/conda-anaconda-tos_1752801584932/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_2d41dz5u60/croot/conda-package-streaming_1751307521903/work
coverage==7.10.4
cryptography==46.0.1
cyclonedx-bom==7.0.0
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1729059153117/work
dparse==0.6.4
filelock==3.19.1
fqdn==1.5.1
frozendict @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/frozendict_1729090952168/work
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isoduration==20.11.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
license-expression==30.4.4
loguru==0.7.3
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mdurl==0.1.2
menuinst @ file:///C:/b/abs_a0aibbe6s6/croot/menuinst_1750861726761/work
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.3
outcome==1.3.0.post0
packageurl-python==0.17.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pbr==7.0.0
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip-tools==7.5.0
pip_audit==2.9.0
platformdirs==4.4.0
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
pre_commit==4.3.0
propcache==0.3.2
psutil==7.1.0
py-serializable==2.1.0
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser==2.23
pydantic==2.11.9
pydantic_core @ file:///C:/b/abs_d0xs0z3erk/croot/pydantic-core_1750754747982/work
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1729039320841/work
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==14.1.0
rpds-py==0.27.0
ruamel.yaml @ file:///C:/b/abs_0dn1nhdq75/croot/ruamel.yaml_1745960327187/work
ruamel.yaml.clib @ file:///C:/b/abs_475flulhy_/croot/ruamel.yaml.clib_1745937166245/work
safety==3.6.2
safety-schemas==0.0.16
schedule==1.2.2
selenium==4.35.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stevedore==5.4.1
style==1.1.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
trelby==2.4.16.2
trio==0.30.0
trio-websocket==0.12.2
truststore @ file:///C:/b/abs_494cm143zh/croot/truststore_1736550137835/work
typer==0.19.2
types-python-dateutil==2.9.0.20250809
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
update==0.0.1
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
webcolors==24.11.1
websocket-client==1.8.0
wheel==0.45.1
win32_setctime==1.2.0
win_inet_pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1729038644686/work
wsproto==1.2.0
wxPython==4.2.4
yamllint==1.37.1
yarl==1.20.1
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
